# contracts/openapi/oceanembed.v1.yaml
openapi: 3.1.0
info:
  title: OceanEmbed REST API
  version: "1.0.0"
  description: >
    OceanEmbed Subsurface Ocean State Reconstruction and Argo Advisor API.
    Provides reconstructed subsurface 3D ocean temperature fields, Argo in-situ validation overlays,
    uncertainty quantification, physical diagnostic layers (OHC, D20, D26), and deployment site recommendations.
servers:
  - url: /v1
    description: Production API Server v1
paths:
...
  /section:
    get:
      summary: Get vertical transect / cross-section
      description: Interpolates subsurface temperature along an arbitrary multi-point polyline path.
      parameters:
        - name: date
          in: query
          required: true
          schema: { type: string, format: date }
        - name: path
          in: query
          required: true
          description: Semicolon-separated lat,lon coordinates e.g. "10.0,60.0;15.0,75.0;20.0,90.0"
          schema: { type: string }
      responses:
        '200':
          description: Vertical transect array [depth, path_points]
          content:
            application/json:
              schema:
                type: object
                properties:
                  date: { type: string }
                  points: { type: array, items: { type: object } }
                  depths: { type: array, items: { type: number } }
                  temperatures: { type: array, items: { type: array, items: { type: number, nullable: true } } }
...
  /heat:
    get:
      summary: Get Ocean Heat Content (OHC) or Isotherm Depth (D26, D20)
      parameters:
        - name: date
          in: query
          required: true
          schema: { type: string, format: date }
        - name: metric
          in: query
          required: true
          schema: { type: string, enum: [ohc, d26, d20] }
      responses:
        '200':
          description: 2D heat metric grid
          content:
            application/json:
              schema:
                type: object
                properties:
                  date: { type: string }
                  metric: { type: string }
                  units: { type: string }
                  grid: { type: array, items: { type: array, items: { type: number, nullable: true } } }
...